First attempt at the pulsemeter
This is basically the finished version of what I envisioned,
except I've kind of just realized it doesn't actually work
all that well and has some pretty serious limitations.

I spent an offensive amount of time on this and I feel like even
more of a jackass than I already knew I was.

I guess it has to be done with timers?  I'm supposing perhaps
alternating between measuring amplitude and just treating it like
an event trigger?
The reason I didn't use timers before was because I thought there
wasn't a really good way to trigger them with enough accuracy
to be useful.  If it were connected to another pin that has
other functionality then there are ways, but I'll have to rethink
a lot here.